Add Electron packaging and CSV fs support
