+++ package.json
@@ -3,7 +3,13 @@
   "version": "1.0.0",
   "description": "Sinóptico de Producto Barack",
   "license": "MIT",
+  "main": "main.js",
   "scripts": {
-    "test": "echo \"No tests available\" && exit 0"
+    "test": "echo \"No tests available\" && exit 0",
+    "start": "electron ."
+  },
+  "devDependencies": {
+    "electron": "^27.1.1",
+    "electron-packager": "^17.1.1"
   }
 }
