+++ preload.js
@@ -1,0 +1,1 @@
+// This file can be used to expose safe APIs to the renderer
