+++ renderer.js
@@ -1,0 +1,562 @@
+    document.addEventListener('DOMContentLoaded', () => {
+      const fs = window.require ? window.require("fs") : null;
+      /* ==================================================
+         1) Mostrar/Ocultar Columnas
+      ================================================== */
+      const toggles = document.querySelectorAll('.toggle-col');
+      toggles.forEach(chk => {
+        chk.addEventListener('change', () => {
+          const colIndex = parseInt(chk.getAttribute('data-colindex'));
+          const checked = chk.checked;
+          // Ocultar o mostrar <th>
+          const th = document.querySelector(`#sinoptico thead th:nth-child(${colIndex + 1})`);
+          if (th) th.style.display = checked ? '' : 'none';
+          // Ocultar o mostrar <td> de esa columna
+          document.querySelectorAll(`#sinoptico tbody tr`).forEach(tr => {
+            const td = tr.querySelector(`td:nth-child(${colIndex + 1})`);
+            if (td) td.style.display = checked ? '' : 'none';
+          });
+        });
+      });
+
+      /* ==================================================
+         2) Funciones de alerta y filtrado
+      ================================================== */
+      function mostrarMensaje(texto, tipo = 'error') {
+        const div = document.getElementById('mensaje');
+        div.textContent = texto;
+        div.style.display = 'block';
+        div.style.backgroundColor = tipo === 'error' ? '#e74c3c' : '#f39c12';
+        div.style.color = '#ffffff';
+      }
+
+      function aplicarFiltro() {
+        const criterio = document.getElementById('filtroInsumo').value.trim().toLowerCase();
+        const incluirAncestros = document.getElementById('chkIncluirAncestros').checked;
+        const mostrar0 = document.getElementById('chkMostrarNivel0').checked;
+        const mostrar1 = document.getElementById('chkMostrarNivel1').checked;
+        const mostrar2 = document.getElementById('chkMostrarNivel2').checked;
+        const mostrar3 = document.getElementById('chkMostrarNivel3').checked;
+
+        const todasFilas = Array.from(document.querySelectorAll('#sinoptico tbody tr'));
+        const mapIdToRow = {};
+        todasFilas.forEach(tr => {
+          const id = tr.getAttribute('data-id');
+          if (id) mapIdToRow[id] = tr;
+        });
+
+        function mostrarAncestros(id) {
+          if (!id) return;
+          const fila = mapIdToRow[id];
+          if (!fila) return;
+          fila.style.display = '';
+          const parentId = fila.getAttribute('data-parent');
+          if (parentId) mostrarAncestros(parentId);
+        }
+
+        if (!criterio) {
+          // Mostrar/ocultar según nivel, sin filtro de texto
+          todasFilas.forEach(tr => {
+            const nivel = tr.classList.contains('nivel-0')
+              ? 0
+              : tr.classList.contains('nivel-1')
+              ? 1
+              : tr.classList.contains('nivel-2')
+              ? 2
+              : 3;
+            const mostrarEste =
+              (nivel === 0 && mostrar0) ||
+              (nivel === 1 && mostrar1) ||
+              (nivel === 2 && mostrar2) ||
+              (nivel === 3 && mostrar3);
+            tr.style.display = mostrarEste ? '' : 'none';
+          });
+          return;
+        }
+
+        // Con texto: ocultar todo y luego mostrar coincidencias + ancestros
+        todasFilas.forEach(tr => (tr.style.display = 'none'));
+        todasFilas.forEach(tr => {
+          const textoFila = tr.textContent.toLowerCase();
+          if (textoFila.includes(criterio)) {
+            const nivel = tr.classList.contains('nivel-0')
+              ? 0
+              : tr.classList.contains('nivel-1')
+              ? 1
+              : tr.classList.contains('nivel-2')
+              ? 2
+              : tr.classList.contains('nivel-3')
+              ? 3
+              : 4;
+            const mostrarEste =
+              (nivel === 0 && mostrar0) ||
+              (nivel === 1 && mostrar1) ||
+              (nivel === 2 && mostrar2) ||
+              (nivel === 3 && mostrar3);
+            if (mostrarEste) tr.style.display = '';
+            if (incluirAncestros) {
+              const parentId = tr.getAttribute('data-parent');
+              mostrarAncestros(parentId);
+            }
+          }
+        });
+
+        // Finalmente ocultar niveles no marcados
+        todasFilas.forEach(tr => {
+          const nivel = tr.classList.contains('nivel-0')
+            ? 0
+            : tr.classList.contains('nivel-1')
+            ? 1
+            : tr.classList.contains('nivel-2')
+            ? 2
+            : tr.classList.contains('nivel-3')
+            ? 3
+            : 4;
+          const mostrarEste =
+            (nivel === 0 && mostrar0) ||
+            (nivel === 1 && mostrar1) ||
+            (nivel === 2 && mostrar2) ||
+            (nivel === 3 && mostrar3);
+          if (!mostrarEste) tr.style.display = 'none';
+        });
+      }
+
+      // Adjuntamos eventos a campos de filtro
+      document.getElementById('filtroInsumo').addEventListener('keyup', aplicarFiltro);
+      document.getElementById('chkIncluirAncestros').addEventListener('change', aplicarFiltro);
+      document.getElementById('chkMostrarNivel0').addEventListener('change', aplicarFiltro);
+      document.getElementById('chkMostrarNivel1').addEventListener('change', aplicarFiltro);
+      document.getElementById('chkMostrarNivel2').addEventListener('change', aplicarFiltro);
+      document.getElementById('chkMostrarNivel3').addEventListener('change', aplicarFiltro);
+
+      /* ==================================================
+         3) Botones Expandir / Colapsar (nodos + recursivo)
+      ================================================== */
+      function showChildren(parentId) {
+        const hijos = document.querySelectorAll(`#sinoptico tbody tr[data-parent="${parentId}"]`);
+        hijos.forEach(hijo => {
+          hijo.style.display = '';
+        });
+      }
+      function hideSubtree(parentId) {
+        const hijos = document.querySelectorAll(`#sinoptico tbody tr[data-parent="${parentId}"]`);
+        hijos.forEach(hijo => {
+          hijo.style.display = 'none';
+          const btn = hijo.querySelector('.toggle-btn');
+          if (btn) {
+            btn.textContent = '+';
+            btn.setAttribute('data-expanded', 'false');
+            btn.setAttribute('aria-expanded', 'false');
+          }
+          const idHijo = hijo.getAttribute('data-id');
+          hideSubtree(idHijo);
+        });
+      }
+      function toggleNodo(btn, parentId) {
+        const expanded = btn.getAttribute('data-expanded') === 'true';
+        if (expanded) {
+          // Colapsar
+          btn.textContent = '+';
+          btn.setAttribute('data-expanded', 'false');
+          btn.setAttribute('aria-expanded', 'false');
+          hideSubtree(parentId);
+        } else {
+          // Expandir
+          btn.textContent = '–';
+          btn.setAttribute('data-expanded', 'true');
+          btn.setAttribute('aria-expanded', 'true');
+          showChildren(parentId);
+        }
+      }
+      function colapsarTodo() {
+        document.querySelectorAll('#sinoptico tbody tr').forEach(tr => {
+          if (!tr.classList.contains('nivel-0')) {
+            tr.style.display = 'none';
+          }
+        });
+        document.querySelectorAll('.toggle-btn').forEach(btn => {
+          if (!btn.classList.contains('hidden')) {
+            btn.textContent = '+';
+            btn.setAttribute('data-expanded', 'false');
+            btn.setAttribute('aria-expanded', 'false');
+          }
+        });
+      }
+      function expandirTodo() {
+        document.querySelectorAll('#sinoptico tbody tr').forEach(tr => {
+          tr.style.display = '';
+        });
+        document.querySelectorAll('.toggle-btn').forEach(btn => {
+          if (!btn.classList.contains('hidden')) {
+            btn.textContent = '–';
+            btn.setAttribute('data-expanded', 'true');
+            btn.setAttribute('aria-expanded', 'true');
+          }
+        });
+      }
+      document.getElementById('expandirTodo').addEventListener('click', expandirTodo);
+      document.getElementById('colapsarTodo').addEventListener('click', colapsarTodo);
+
+      /* ==================================================
+         4) Exportar a Excel (solo filas visibles)
+      ================================================== */
+      document.getElementById('btnExcel').addEventListener('click', () => {
+        const filasVisibles = [];
+        // Encabezados (solo los <th> visibles)
+        const encabezados = Array.from(
+          document.querySelectorAll('#sinoptico thead th')
+        )
+          .filter(th => th.style.display !== 'none')
+          .map(th => th.textContent.trim());
+        filasVisibles.push(encabezados);
+
+        // Filas visibles
+        document.querySelectorAll('#sinoptico tbody tr').forEach(tr => {
+          if (tr.style.display === '') {
+            // Solo las celdas <td> que correspondan a encabezados visibles
+            const celdas = Array.from(tr.querySelectorAll('td'))
+              .filter((td, idx) => {
+                const th = document.querySelector(`#sinoptico thead th:nth-child(${idx + 1})`);
+                return th && th.style.display !== 'none';
+              })
+              .map(td => td.textContent.trim());
+            filasVisibles.push(celdas);
+          }
+        });
+
+        const wb = XLSX.utils.book_new();
+        const ws = XLSX.utils.aoa_to_sheet(filasVisibles);
+        XLSX.utils.book_append_sheet(wb, ws, 'Sinoptico');
+        XLSX.writeFile(wb, 'sinoptico.xlsx');
+      });
+
+      /* ==================================================
+         5) Cargar CSV y construir la tabla jerárquica
+         + Recarga automática cada 30 segundos
+      ================================================== */
+      function procesarDatos(datosOriginal, expandedIds) {
+        if (!datosOriginal.length) {
+          mostrarMensaje('sinoptico.csv se cargó, pero está vacío.');
+          return;
+        }
+        const tbody = document.querySelector('#sinoptico tbody');
+        tbody.textContent = '';
+        const clientesMap = {};
+        datosOriginal.forEach(fila => {
+          const cliente = fila.Cliente;
+          if (!clientesMap[cliente]) {
+            clientesMap[cliente] = {
+              ID: 'cli-' + cliente,
+              ParentID: '',
+              Tipo: 'Cliente',
+              Secuencia: '0',
+              'Ensamble / Subensamble / Operación': cliente,
+              Cliente: cliente,
+              Vehículo: '',
+              RefInterno: '',
+              Versión: '',
+              Imagen: '',
+              'Pzas/h': '',
+              Unidad: '',
+              Sourcing: '',
+              Código: ''
+            };
+          }
+        });
+        const datosMod = [];
+        Object.values(clientesMap).forEach(nodoCli => datosMod.push(nodoCli));
+        datosOriginal.forEach(fila => {
+          if (!fila.ParentID || fila.ParentID.trim() === '') {
+            fila.ParentID = 'cli-' + fila.Cliente;
+          }
+          datosMod.push(fila);
+        });
+        construirSinoptico(datosMod);
+        setTimeout(ajustarIndentacion, 20);
+        setTimeout(() => {
+          expandedIds.forEach(id => {
+            showChildren(id);
+            const btn = document.querySelector(`#sinoptico tbody tr[data-id="${id}"] .toggle-btn`);
+            if (btn) {
+              btn.textContent = '–';
+              btn.setAttribute('data-expanded', 'true');
+              btn.setAttribute('aria-expanded', 'true');
+            }
+          });
+          aplicarFiltro();
+        }, 40);
+      }
+
+      function loadData() {
+        // Guardar IDs de filas actualmente expandidas antes de limpiar
+        const expandedIds = Array.from(
+          document.querySelectorAll('#sinoptico tbody .toggle-btn[data-expanded="true"]')
+        ).map(btn => btn.closest('tr').getAttribute('data-id'));
+
+
+        if (fs) {
+          try {
+            const csvText = fs.readFileSync('sinoptico.csv', 'utf8');
+            const results = Papa.parse(csvText, {
+              header: true,
+              skipEmptyLines: true,
+              delimiter: ';'
+            });
+            if (results.errors.length) {
+              console.error('Errores al parsear CSV:', results.errors);
+              mostrarMensaje('Hubo errores al parsear sinoptico.csv.');
+              return;
+            }
+            const datosOriginal = results.data;
+            procesarDatos(datosOriginal, expandedIds);
+          } catch (err) {
+            console.error('Error al leer sinoptico.csv:', err);
+            mostrarMensaje('No se pudo cargar sinoptico.csv (404 o problema de ruta).');
+          }
+          return;
+        }
+
+        Papa.parse('sinoptico.csv', {
+          download: true,
+          header: true,
+          skipEmptyLines: true,
+          delimiter: ';',
+          complete(results) {
+            if (results.errors.length) {
+              console.error('Errores al parsear CSV:', results.errors);
+              mostrarMensaje('Hubo errores al parsear sinoptico.csv.');
+              return;
+            }
+            const datosOriginal = results.data;
+              procesarDatos(datosOriginal, expandedIds);
+          },
+          error(err) {
+            console.error('Error al leer sinoptico.csv:', err);
+            mostrarMensaje('No se pudo cargar sinoptico.csv (404 o problema de ruta).');
+          }
+        });
+      }
+
+      // Llamo inmediatamente a loadData()
+      loadData();
+      // Luego, cada 30 segundos, recargo el CSV automáticamente
+      setInterval(loadData, 30000);
+
+
+      // Función recursiva para poblar la tabla desde datos con "ID" y "ParentID"
+      function construirSinoptico(datos) {
+        // a) Agrupar por ParentID
+        const agrupadoPorPadre = {};
+        datos.forEach(fila => {
+          if (!fila.ID) return;
+          const padre = (fila.ParentID || '').toString().trim();
+          if (!agrupadoPorPadre[padre]) agrupadoPorPadre[padre] = [];
+          agrupadoPorPadre[padre].push(fila);
+        });
+        // b) Ordenar cada grupo según Secuencia
+        Object.keys(agrupadoPorPadre).forEach(key => {
+          agrupadoPorPadre[key].sort((a, b) => {
+            const sa = parseInt(a.Secuencia) || 0;
+            const sb = parseInt(b.Secuencia) || 0;
+            return sa - sb;
+          });
+        });
+
+        // c) Dibujar filas recursivamente
+        function dibujarNodos(parentID, nivel) {
+          const hijos = agrupadoPorPadre[parentID] || [];
+          hijos.forEach(fila => {
+            const tr = document.createElement('tr');
+            tr.setAttribute('data-id', fila.ID);
+            tr.setAttribute('data-parent', (fila.ParentID || '').toString().trim());
+
+            // Color según Tipo
+            const tipoStr = (fila.Tipo || '').toString().trim().toLowerCase();
+            if (tipoStr === 'cliente') tr.classList.add('fila-cliente');
+            else if (tipoStr === 'pieza final') tr.classList.add('fila-pieza');
+            else if (tipoStr === 'subensamble') tr.classList.add('fila-subens');
+            else tr.classList.add('fila-oper');
+
+            // Clase de nivel (0..6)
+            tr.classList.add(`nivel-${nivel}`);
+
+            // Celda “Item” con botón +/–, flecha y texto
+            const tdItem = document.createElement('td');
+            // Permitir que el texto se muestre completo
+            tdItem.style.whiteSpace = 'normal';
+            tdItem.style.overflow = 'visible';
+            tdItem.style.textOverflow = 'unset';
+
+            // ¿Tiene hijos este nodo?
+            const tieneHijos =
+              Array.isArray(agrupadoPorPadre[fila.ID]) &&
+              agrupadoPorPadre[fila.ID].length > 0;
+            const btnToggle = document.createElement('button');
+            btnToggle.type = 'button';
+            btnToggle.classList.add('toggle-btn');
+            if (!tieneHijos) {
+              btnToggle.classList.add('hidden');
+            }
+            btnToggle.textContent = '+'; // inicia colapsado
+            btnToggle.setAttribute('data-expanded', 'false');
+            btnToggle.setAttribute('aria-expanded', 'false');
+            btnToggle.setAttribute('aria-label', 'Expand/Collapse');
+            // Evento onclick
+            btnToggle.addEventListener('click', () => {
+              toggleNodo(btnToggle, fila.ID);
+            });
+            tdItem.appendChild(btnToggle);
+
+            let nombreItem = (fila['Ensamble / Subensamble / Operación'] || '').toString().trim();
+            if (
+              nivel === 1 &&
+              fila.Tipo.toString().trim().toLowerCase() === 'pieza final'
+            ) {
+              nombreItem = 'Barack ' + nombreItem;
+            }
+
+            if (nivel === 0) {
+              // Nivel 0 (Cliente): texto en negrita, sin flecha
+              const spanText = document.createElement('span');
+              spanText.classList.add('item-text');
+              const strong = document.createElement('strong');
+              strong.textContent = nombreItem;
+              spanText.appendChild(strong);
+              tdItem.appendChild(spanText);
+            } else {
+              // Nivel ≥1: flecha + texto
+              const spanArrow = document.createElement('span');
+              spanArrow.classList.add(`arrow-nivel-${nivel}`);
+              const spanText = document.createElement('span');
+              spanText.classList.add('item-text');
+              spanText.textContent = nombreItem;
+              tdItem.appendChild(spanArrow);
+              tdItem.appendChild(spanText);
+            }
+            tr.appendChild(tdItem);
+
+            // Columnas fijas: Cliente, Vehículo, RefInterno, Versión
+            ['Cliente', 'Vehículo', 'RefInterno', 'Versión'].forEach(campo => {
+              const td = document.createElement('td');
+              td.textContent = fila[campo] ? fila[campo].toString().trim() : '';
+              td.style.textAlign = 'center';
+              tr.appendChild(td);
+            });
+
+            // Columna Imagen (miniatura)
+            const tdImagen = document.createElement('td');
+            const nombreImg = (fila['Imagen'] || '').toString().trim();
+            if (nombreImg) {
+              const img = document.createElement('img');
+              img.src = `images/${nombreImg}`;
+              img.alt = nombreItem;
+              img.classList.add('img-product');
+              tdImagen.appendChild(img);
+            }
+            tdImagen.style.textAlign = 'center';
+            tr.appendChild(tdImagen);
+
+            // Columnas: Pzas/h, Unidad, Sourcing, Código
+            ['Pzas/h', 'Unidad', 'Sourcing', 'Código'].forEach(campo => {
+              const td = document.createElement('td');
+              td.textContent = fila[campo] ? fila[campo].toString().trim() : '';
+              td.style.textAlign = 'center';
+              tr.appendChild(td);
+            });
+
+            document.querySelector('#sinoptico tbody').appendChild(tr);
+            dibujarNodos(fila.ID.toString().trim(), nivel + 1);
+          });
+        }
+        dibujarNodos('', 0);
+      }
+
+      /* ==================================================
+         6) Ajustar indentación dinámica (nivel ≥ 1)
+      ================================================== */
+      function ajustarIndentacion() {
+        const todasFilas = Array.from(document.querySelectorAll('#sinoptico tbody tr'));
+        const mapIdToRow = {};
+        todasFilas.forEach(tr => {
+          const id = tr.getAttribute('data-id');
+          if (id) mapIdToRow[id] = tr;
+        });
+
+        // Cache de indentaciones
+        const indentCache = {};
+
+        function obtenerIndent(tr) {
+          const nivel = tr.classList.contains('nivel-0')
+            ? 0
+            : tr.classList.contains('nivel-1')
+            ? 1
+            : tr.classList.contains('nivel-2')
+            ? 2
+            : tr.classList.contains('nivel-3')
+            ? 3
+            : tr.classList.contains('nivel-4')
+            ? 4
+            : tr.classList.contains('nivel-5')
+            ? 5
+            : 6;
+          if (nivel === 0) {
+            return 8; // padding-left fijo para nivel 0
+          }
+          const id = tr.getAttribute('data-id');
+          if (indentCache[id] !== undefined) {
+            return indentCache[id];
+          }
+          // Obtener fila padre
+          const parentId = tr.getAttribute('data-parent');
+          const filaPadre = mapIdToRow[parentId];
+          if (!filaPadre) {
+            indentCache[id] = 8;
+            return 8;
+          }
+          // Indent del padre
+          const indentPadre = obtenerIndent(filaPadre);
+          // Medir anchos: botón toggle + flecha + texto del padre
+          const tdPadre = filaPadre.querySelector('td:first-child');
+          let anchoToggle = 0,
+            anchoArrow = 0,
+            anchoText = 0;
+          const spanToggle = tdPadre.querySelector('.toggle-btn:not(.hidden)');
+          if (spanToggle)
+            anchoToggle =
+              spanToggle.offsetWidth + parseInt(getComputedStyle(spanToggle).marginRight);
+          const spanArrowPadre = tdPadre.querySelector(`[class^="arrow-nivel-"]`);
+          if (spanArrowPadre)
+            anchoArrow =
+              spanArrowPadre.offsetWidth +
+              parseInt(getComputedStyle(spanArrowPadre).marginRight);
+          const spanTextPadre = tdPadre.querySelector('.item-text');
+          if (spanTextPadre) anchoText = spanTextPadre.offsetWidth;
+
+          const margenExtra = 4; // sólo 4px en lugar de 8px
+          const indentActual = indentPadre + anchoToggle + anchoArrow + anchoText + margenExtra;
+          indentCache[id] = indentActual;
+          return indentActual;
+        }
+
+        todasFilas.forEach(tr => {
+          const nivel = tr.classList.contains('nivel-0')
+            ? 0
+            : tr.classList.contains('nivel-1')
+            ? 1
+            : tr.classList.contains('nivel-2')
+            ? 2
+            : tr.classList.contains('nivel-3')
+            ? 3
+            : tr.classList.contains('nivel-4')
+            ? 4
+            : tr.classList.contains('nivel-5')
+            ? 5
+            : 6;
+          if (nivel === 0) return;
+          const indent = obtenerIndent(tr);
+          const tdHijo = tr.querySelector('td:first-child');
+          tdHijo.style.paddingLeft = indent + 'px';
+        });
+      }
+
+    }); // FIN DOMContentLoaded
