+++ sinoptico.html
@@ -96,6 +96,6 @@
   </div>
   <a href="index.html" class="home-button">Inicio</a>
 
-  <script src="main.js" defer></script>
+  <script src="renderer.js" defer></script>
 </body>
 </html>
